Fix logic for import session drawer
- Properly re-fetch session data
- Rendering improvements
